update max upload size, thresholding, and normalize

+++ build/utils/model.js
@@ -100,18 +100,25 @@ class ModelUtils {
                 },
                 data: modelInfo,
             });
-            yield __1.prisma.modelClass.deleteMany({
+            const oldClassModel = yield __1.prisma.modelClass.findMany({
                 where: {
                     mlmodel_id: model_id,
                 },
             });
-            for (let i = 0; i < classAmount; i++) {
-                yield __1.prisma.modelClass.create({
-                    data: {
+            if (classAmount != oldClassModel.length) {
+                yield __1.prisma.modelClass.deleteMany({
+                    where: {
                         mlmodel_id: model_id,
-                        index: i,
                     },
                 });
+                for (let i = 0; i < classAmount; i++) {
+                    yield __1.prisma.modelClass.create({
+                        data: {
+                            mlmodel_id: model_id,
+                            index: i,
+                        },
+                    });
+                }
             }
         });
     }
